Create Day 50 Challenge
Day 50! Boy, you are smashing this 100 days! 🎊 To celebrate, here's a project for you.

Your idea storage system should:

1. Prompt the user to add an idea, or load a random one.
2. Choosing 'Add an idea' results in:
i. A prompt for the user to input their idea.
ii. Add the idea to a text file called 'my.ideas'.
iii. Add the idea in append mode, with every new idea on a brand 
    new line.
3. Choosing 'Load random' results in:
i. Load the list of ideas.
ii. Choose one at random.
iii. Display it on screen for a few seconds.
iv. Return to the menu.