Day 59 Challenge
👉 Day 59 Challenge
You're going to write a program that checks a string to see if it is a palindrome.

Yes. We know that Python has the built in function string.reverse() that you could use.

Zero points for that today, we want you to think hard and utilize your skills in:

i. recursion
ii. string slicing
iii. looping

Your program should:

1. Prompt the user to input a word.
2. Analyze the word to see if it is a palindrome.
3. Output a relevant 'yes/no message.