Day 87 Challenge
👉 Day 87 Challenge
Today's challenge is to edit your login page for blog engine system from yesterday. You should end up with significantly less code than you had before.

Your program should:

Change the login button to forward the user to the edit page.
On the edit page, check that it's you. If not, kick the user back to the main page.