Create Day 89 Challenge
👉 Day 89 Challenge
Today's challenge is to build a community chat app.

Create a login screen. You can use the simple or complex version of Replit authentication. It's up to you.

Once authenticated, users will be directed to a chat room where:

They can see the last five messages on screen.
They can add a new message.
Each message should be displayed with the username and profile pic of the poster.

Allow only your username to be an admin. This should:

Give you and you alone access to an 'admin' button.
Give you the power to delete any message from the board with a delete button.
Don't worry about auto refreshing the chat just yet. We're not getting into Javascript in this 100 days.